@@ -12,28 +12,30 @@
     "studio": "npx prisma studio"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.282.0",
-    "@headlessui/react": "^1.7.11",
+    "@aws-sdk/client-s3": "^3.288.0",
+    "@headlessui/react": "^1.7.13",
     "@heroicons/react": "^2.0.16",
     "@next-auth/prisma-adapter": "^1.0.5",
-    "@prisma/client": "^4.10.0",
+    "@prisma/client": "^4.11.0",
     "clsx": "^1.2.1",
-    "eslint": "^8.29.0",
-    "next": "^13.2.3",
-    "next-auth": "^4.19.2",
-    "prisma": "^4.10.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scroll-parallax": "^3.3.2"
+    "eslint": "^8.35.0",
+    "next": "13.2.4",
+    "next-auth": "^4.20.1",
+    "prisma": "^4.11.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "server-only": "^0.0.1"
   },
   "devDependencies": {
-    "@types/node": "18.14.6",
+    "@types/node": "^18.15.0",
     "@types/react": "^18.0.28",
-    "autoprefixer": "^10.4.13",
-    "eslint-config-next": "^13.2.3",
-    "eslint-config-prettier": "^8.5.0",
-    "postcss": "^8.4.20",
-    "prettier": "2.8.1",
-    "tailwindcss": "^3.2.4"
+    "autoprefixer": "^10.4.14",
+    "eslint-config-next": "^13.2.4",
+    "eslint-config-prettier": "^8.7.0",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.4",
+    "prettier-plugin-tailwindcss": "^0.2.4",
+    "tailwindcss": "^3.2.7",
+    "typescript": "^4.9.5"
   }
 }
