@@ -6,7 +6,8 @@ const nextConfig = {
   },
   experimental: {
     appDir: true,
-  }
+    typedRoutes: true,
+  },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
