@@ -1,4 +1,5 @@
-import React, { useEffect, useMemo, useState } from "react";
+"use client";
+import { useEffect, useMemo, useState } from "react";
 import Shape from "./Shape";
 
 const selectRandom = (arr: any[]) => arr[Math.floor(Math.random() * arr.length)];
